housekeeping: refactor task detail screen into separate files
