+++ lib/presentation/screens/task_detail_screen/task_detail_screen.dart
@@ -2,12 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 
-import '../../components/templates/detail_screen.dart';
-import '../../config/route/routes.dart';
-import '../../config/themes/colours.dart';
-import '../../domain/usecases/task_usecase.dart';
-import '../models/task.dart';
-import '../models/user.dart';
+import '../../../components/templates/detail_screen.dart';
+import '../../../config/route/routes.dart';
+import '../../../config/themes/colours.dart';
+import '../../../domain/usecases/task_usecase.dart';
+import '../../models/task.dart';
+import '../../models/user.dart';
+import 'detail_item.dart';
 
 const _kSpacingBetweenItems = SizedBox(height: 33.0);
 
@@ -59,18 +60,18 @@ class _TaskDetailScreenState extends State<TaskDetailScreen> {
             _kSpacingBetweenItems,
             // notes
             _task.notes != null && _task.notes!.isNotEmpty
-                ? _DetailItem(content: _task.notes!, label: 'Notes')
+                ? DetailItem(content: _task.notes!, label: 'Notes')
                 : Container(),
             _task.notes != null ? _kSpacingBetweenItems : Container(),
             // due date
-            _DetailItem(content: '${_task.due ?? '-'}', label: 'Due Date'),
+            DetailItem(content: '${_task.due ?? '-'}', label: 'Due Date'),
             _kSpacingBetweenItems,
             // available
-            _DetailItem(
+            DetailItem(
                 content: '${_task.completed ?? '-'}', label: 'Available'),
             _kSpacingBetweenItems,
             // rewards
-            _DetailItem(
+            DetailItem(
               content: _task.reward != null ? '${_task.reward} Cryois' : '-',
               label: 'Rewards',
             ),
@@ -90,41 +91,6 @@ class _TaskDetailScreenState extends State<TaskDetailScreen> {
           taskId: widget.id,
         );
       },
-    );
-  }
-}
-
-/// Item for [TaskDetailScreen].
-class _DetailItem extends StatelessWidget {
-  const _DetailItem({Key? key, required this.content, required this.label})
-      : super(key: key);
-
-  final String content;
-  final String label;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Text(
-          label,
-          style: GoogleFonts.jetBrainsMono(
-            color: Colours.darkText,
-            fontSize: 17.0,
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-        const SizedBox(height: 7.0),
-        Text(
-          content,
-          style: GoogleFonts.jetBrainsMono(
-            color: Colours.text,
-            fontSize: 15.0,
-            fontWeight: FontWeight.w300,
-          ),
-        ),
-      ],
     );
   }
 }
